% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_params.R
\name{gamma_params}
\alias{gamma_params}
\title{gamma_params}
\usage{
gamma_params(norm_mean, norm_sd)
}
\arguments{
\item{norm_mean}{The mean value of the associated normal distribution - REQUIRED.}

\item{norm_sd}{The standard deviation of the associated normal distribution - REQUIRED.}
}
\description{
Obtain the parameters from a Gamma distribution based on the equivalent parameters from a normal distribution.
Returns a list of 'loc' and 'scale'.
}
\examples{
gamma_parameters <- gamma_params(norm_mean = 1, norm_sd = 1)
gamma_scale <- gamma_parameters$scale
}
\keyword{Distribution}
\keyword{Gamma,}
\keyword{Parameters}
