% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_exp_quad_cov.R
\name{gen_exp_quad_cov}
\alias{gen_exp_quad_cov}
\title{gen_exp_quad_cov}
\usage{
gen_exp_quad_cov(
  n_dims,
  alpha_sq = 1,
  rho_sq = 3,
  sigma_sq = 1,
  tidy = FALSE,
  longer = FALSE
)
}
\arguments{
\item{n_dims}{The number of dimensions of the covariance matrix - REQUIRED.}

\item{alpha_sq}{The maximum co-variance - defaults to 1.}

\item{rho_sq}{The length scale - defaults to 3.}

\item{sigma_sq}{The (within-dimensional) variance - defaults to 1.}

\item{tidy}{Logical: If 'TRUE', returns a tibble, if 'FALSE'  returns a matrix - defaults to FALSE.}

\item{longer}{Logical: If 'TRUE', returns a long-format tibble, if 'FALSE'  returns a wide-format tibble (requires tidy = TRUE) - defaults to FALSE.}
}
\description{
Generate an n-dimensional covariance matrix, based on the exponential-quadratic covariance function.
Returns either a matrix (tidy = FALSE), a wide tibble (tidy = TRUE, longer = FALSE) or a long tibble (tidy = TRUE, longer = TRUE).
}
\examples{
ten_dim_cov_mat <- cov_mat <- gen_exp_quad_cov(n_dims = 10, tidy = FALSE)
ten_dim_cov_tibble_long <- gen_exp_quad_cov(n_dims = 10, tidy = TRUE, longer = TRUE)
}
\keyword{Covariance}
\keyword{Exponentiated-Quadratic,}
