% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TM_squad.R
\name{get_TM_squad}
\alias{get_TM_squad}
\title{get_TM_squad}
\usage{
get_TM_squad(url, raw = FALSE)
}
\arguments{
\item{url}{A valid ()detailed) transfermarkt url. Navigate to a club page, click on the 'detailed' tab, and then paste the url - REQUIRED.}

\item{raw}{Logical: if TRUE, returns the unprocessed table. if FALSE, returns a processed tibble - defaults to FALSE}
}
\description{
A webscraping function for the website 'https://www.transfermarkt.com/'.
Returns a preprocessed tibble of any squad
}
\examples{
Juventus squad <- get_TM_squad(url <-  https://www.transfermarkt.co.uk/juventus-fc/kader/verein/506/saison_id/2020/plus/1)
Blackburn_Rovers_squad <- get_TM_squad(url <- 'https://www.transfermarkt.co.uk/blackburn-rovers/kader/verein/164/saison_id/2020/plus/1')
}
