% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm_params.R
\name{lnorm_params}
\alias{lnorm_params}
\title{lnorm_params}
\usage{
lnorm_params(norm_mean, norm_sd)
}
\arguments{
\item{norm_mean}{The mean value of the associated normal distribution - REQUIRED.}

\item{norm_sd}{The standard deviation of the associated normal distribution - REQUIRED.}
}
\description{
Obtain the parameters from a lognormal distribution based on the equivalent parameters from a normal distribution.
Returns a list of 'meanlog' and 'sdlog' (equivalent to the arguments from the lnorm family of functions).
}
\examples{
lognormal_parameters <- lnorm_params(norm_mean = 1, norm_sd = 1)
lognormal_mean <- lognormal_parameters$meanlog
}
\keyword{Distribution}
\keyword{Lognormal,}
\keyword{Parameters}
