% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metropolis_RW.R
\name{Metropolis_RW}
\alias{Metropolis_RW}
\title{Metropolis_RW}
\usage{
Metropolis_RW(
  n_dims,
  iterations,
  mu_prior,
  var_prior,
  mu_lik,
  var_lik,
  n_chains = 4,
  n_burn_in = iterations/2,
  x_initials,
  prop_var = 1,
  annealing = 1,
  MCMC_seed = 1008,
  thinning = 1
)
}
\arguments{
\item{iterations}{The total number of samples, including burn-in, to draw from each chain - REQUIRED.}

\item{mu_prior}{An n-dimesional vector of mean values for the prior model - REQUIRED.}

\item{var_prior}{An n-dimesional covriance matrix for the prior model - REQUIRED.}

\item{mu_lik}{An n-dimesional vector of mean values for the likelihood - REQUIRED.}

\item{var_lik}{An n-dimesional covriance matrix for the likelihood - REQUIRED.}

\item{n_chains}{The number of separate chains (with their own initial co-ordinates) to run - defaults to 4 chains.}

\item{n_burn_in}{The number of samples to discard as part f the burn in phase - defaults to half of all samples.}

\item{x_initials}{An n-dimensional vector of initial values for each chain - defaults to a sequence between +/- n_dims}

\item{prop_var}{The variance of the n-dimensional symmetric proposal distribution. A tuning parameter - defaults to 1}

\item{annealing}{A multiplicative simulated annealing parameter - defaults to 1 (equivalent of no annealing)}

\item{MCMC_seed}{A random seed value for reproducable analysis - defaults to 1008}

\item{thinning}{A parameter determining how many successive iterations should be disregarding to mitigate excessive auto-correlation - defaults to 1}
}
\description{
A Metroplis random walk Markov Chain Monte Carlo sampler to draw from an n-dimensional posterior distribution defined with Gaussian priors and likelihoods
}
\examples{
theme_ddf_light()
}
