% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metropolis_traceplot.R
\name{Metropolis_traceplot}
\alias{Metropolis_traceplot}
\title{Metropolis_traceplot}
\usage{
Metropolis_traceplot(mcmc_data, include_burn_in = FALSE, alpha = 0.4)
}
\arguments{
\item{mcmc_data}{Data of the form produced by DomDF::Metroplois_RW() - REQUIRED.}

\item{include_burn_in}{Whether burn-in samples should be included in the plot. A vertical line is added to show this division when set to TRUE. - Defaults to FALSE.}

\item{alpha}{Transparancy of each chain in the plot - Defaults to 0.4.}
}
\description{
Returns a ggplot object of a traceplot of samples from a Metroplis random walk Markov Chain Monte Carlo sampler (see DomDF::Metropolis_RW())
}
\examples{
Metropolis_traceplot(mcmc_data = four_dim_mcmc_samples)
Metropolis_traceplot(mcmc_data = one_dim_mcmc_samples)

}
