% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_pred_app.R
\name{prior_pred_app}
\alias{prior_pred_app}
\title{Prior predictive sampling Shiny app}
\usage{
prior_pred_app()
}
\description{
Opens an interactive Shiny web app in your browser, which allows visualisation of prior models on the outcome scale for Bayesian GLMs
}
\examples{
prior_pred_app() # Opens the interactive Shiny app

}
\keyword{Bayesian,}
\keyword{Priors}
\keyword{Shiny,}
