% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ddf_dark.R
\name{theme_ddf_dark}
\alias{theme_ddf_dark}
\title{theme_ddf_dark}
\usage{
theme_ddf_dark(
  base_size = 12,
  base_family = extrafont::choose_font(fonts = "Bahnschrift")
)
}
\arguments{
\item{base_size}{The font size to display. Defaults to 12.}

\item{base_family}{The font to display. Defaults to 'Bahnschrift' - a font in the extrafont library.}
}
\description{
A pleasing dark theme for ggplot graphics.
}
\examples{
theme_ddf_dark()
}
