% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mcmc_draws.R
\name{tidy_mcmc_draws}
\alias{tidy_mcmc_draws}
\title{tidy_mcmc_draws}
\usage{
tidy_mcmc_draws(cmdstan_fit, params = "all_params")
}
\arguments{
\item{cmdstan_fit}{A CmdStanMCMC object. These are generated using the $sample method from a CmdStanModel. REQUIRED}

\item{params}{A vector of characters containing the parameters to be extracted. Defaults to 'all_params', which returns all parameters.}
}
\description{
Obtain MCMC samples from a CmdStanR model, formatted for ggplot2 plotting.
Returns a a long, tidy tibble.
}
\examples{
mcmc_df <- tidy_mcmc_draws(cmdstan_fit = Bayesian_model_fit)
reduced_mcmc_df <- tidy_mcmc_draws(cmdstan_fit = Bayesian_model_fit, params = c('A', 'm')) #Where 'A' and 'm' are parameters in the Stan program.
}
\keyword{CmdStanR,}
\keyword{MCMC,}
\keyword{Tibble}
\keyword{Tidy,}
