% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mcmc_gq.R
\name{tidy_mcmc_gq}
\alias{tidy_mcmc_gq}
\title{tidy_mcmc_draws}
\usage{
tidy_mcmc_gq(cmdstan_gq, params = "all_params")
}
\arguments{
\item{cmdstan_gq}{A CmdStan generated quantities object. These are generated using the $generate_quantities method from a CmdStanModel. REQUIRED}

\item{params}{A vector of characters containing the parameters to be extracted. Defaults to 'all_params', which returns all parameters.}
}
\description{
Obtain MCMC samples from a CmdStanR model, formatted for ggplot2 plotting.
Returns a a long, tidy tibble.
}
\examples{
gq_df <- tidy_mcmc_gq(cmdstan_gq = Bayesian_model_gq)
reduced_mcmc_gq <- tidy_mcmc_gq(cmdstan_gq = Bayesian_model_gq, params = c('A', 'm')) # Where 'A' and 'm' are parameters in the Stan program.
}
\keyword{CmdStanR,}
\keyword{MCMC,}
\keyword{Tibble}
\keyword{Tidy,}
