% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metropolis_traceplot.R
\name{Metropolis_traceplot}
\alias{Metropolis_traceplot}
\title{Metropolis_traceplot}
\usage{
Metropolis_traceplot(
  Metropolis_RW_dataframe,
  include_burn_in = FALSE,
  alpha = 0.4
)
}
\arguments{
\item{Metropolis_RW_dataframe}{A dataframe produced by DomDF::Metroplois_RW() - REQUIRED.}

\item{include_burn_in}{Whether burn-in samples should be included in the plot - Defaults to FALSE.}

\item{alpha}{Transparancy of each chain in the plot - Defaults to 0.4.}
}
\description{
A traceplot of samples from a Metroplis random walk Markov Chain Monte Carlo sampler (see DomDF::Metropolis_RW())
}
\examples{
Metropolis_traceplot(Metropolis_RW_dataframe = four_dim_mcmc_df)
Metropolis_traceplot(Metropolis_RW_dataframe = one_dim_mcmc_df)

}
